+++ packages/shadcn/src/utils/transformers/transform-rsc.test.ts
@@ -1,0 +1,113 @@
+import type { Config } from "@/src/utils/get-config"
+import { expect, it } from "vitest"
+
+import { transform } from "."
+
+it("transform rsc", async () => {
+  expect(
+    await transform({
+      filename: "test.ts",
+      raw: `import * as React from "react"
+import { Foo } from "bar"
+    `,
+      config: {
+        tsx: true,
+        rsc: true,
+      } as Config,
+    })
+  ).toMatchSnapshot()
+
+  expect(
+    await transform({
+      filename: "test.ts",
+      raw: `"use client"
+
+      import * as React from "react"
+import { Foo } from "bar"
+    `,
+      config: {
+        tsx: true,
+        rsc: true,
+      } as Config,
+    })
+  ).toMatchSnapshot()
+
+  expect(
+    await transform({
+      filename: "test.ts",
+      raw: `"use client"
+
+      import * as React from "react"
+import { Foo } from "bar"
+    `,
+      config: {
+        tsx: true,
+        rsc: false,
+      } as Config,
+    })
+  ).toMatchSnapshot()
+
+  expect(
+    await transform({
+      filename: "test.ts",
+      raw: `"use foo"
+
+      import * as React from "react"
+import { Foo } from "bar"
+
+"use client"
+    `,
+      config: {
+        tsx: true,
+        rsc: false,
+      } as Config,
+    })
+  ).toMatchSnapshot()
+
+  expect(
+    await transform({
+      filename: "test.ts",
+      raw: `'use client'
+
+      import * as React from 'react'
+import { Foo } from 'bar'
+    `,
+      config: {
+        tsx: true,
+        rsc: true,
+      } as Config,
+    })
+  ).toMatchSnapshot()
+
+  expect(
+    await transform({
+      filename: "test.ts",
+      raw: `'use client'
+
+      import * as React from 'react'
+import { Foo } from 'bar'
+    `,
+      config: {
+        tsx: true,
+        rsc: false,
+      } as Config,
+    })
+  ).toMatchSnapshot()
+
+  expect(
+    await transform({
+      filename: "test.ts",
+      raw: `'use foo'
+
+      import * as React from 'react'
+import { Foo } from 'bar'
+
+'use client'
+    `,
+      config: {
+        tsx: true,
+        rsc: false,
+      } as Config,
+    })
+  ).toMatchSnapshot()
+})